branched off deleting / saving js script to its own file
